requirements/requirements_armv7l.txt
python-statsd==1.6.0
PyYAML==3.10
Adafruit-BBIO==0.0.9
celery==3.1.7
flower==0.6.0
pytz==2013.9
redis==2.9.0
python-dateutil==2.2